Move CellData into new file
