+++ src/InputHandler.ts
@@ -13,10 +13,11 @@ import { IDisposable } from 'xterm';
 import { Disposable } from 'common/Lifecycle';
 import { concat } from 'common/TypedArrayUtils';
 import { StringToUtf32, stringFromCodePoint, utf32ToString, Utf8ToUtf32 } from 'common/input/TextDecoder';
-import { CellData, Attributes, FgFlags, BgFlags, AttributeData, DEFAULT_ATTR_DATA } from 'common/buffer/BufferLine';
+import { Attributes, FgFlags, BgFlags, AttributeData, DEFAULT_ATTR_DATA } from 'common/buffer/BufferLine';
 import { EventEmitter, IEvent } from 'common/EventEmitter';
 import { IParsingState, IDcsHandler, IEscapeSequenceParser } from 'common/parser/Types';
-import { NULL_CELL_CODE, NULL_CELL_WIDTH } from '../out/common/buffer/Constants';
+import { NULL_CELL_CODE, NULL_CELL_WIDTH } from 'common/buffer/Constants';
+import { CellData } from 'common/buffer/CellData';
 
 /**
  * Map collect to glevel. Used in `selectCharset`.
