+++ src/common/buffer/BufferLine.test.ts
@@ -4,7 +4,8 @@
  */
 import * as chai from 'chai';
 import { NULL_CELL_CHAR, NULL_CELL_WIDTH, NULL_CELL_CODE, DEFAULT_ATTR } from 'common/buffer/Constants';
-import { BufferLine, CellData, Content } from 'common/buffer//BufferLine';
+import { BufferLine, Content } from 'common/buffer//BufferLine';
+import { CellData } from 'common/buffer/CellData';
 import { CharData, IBufferLine } from '../Types';
 
 class TestBufferLine extends BufferLine {
