+++ src/renderer/CursorRenderLayer.ts
@@ -7,7 +7,7 @@ import { IRenderDimensions } from 'browser/renderer/Types';
 import { BaseRenderLayer } from './BaseRenderLayer';
 import { ITerminal } from '../Types';
 import { ICellData } from 'common/Types';
-import { CellData } from 'common/buffer/BufferLine';
+import { CellData } from 'common/buffer/CellData';
 import { IColorSet } from 'browser/Types';
 
 interface ICursorState {
