+++ src/renderer/TextRenderLayer.ts
@@ -9,10 +9,11 @@ import { ITerminal } from '../Types';
 import { CharData, ICellData } from 'common/Types';
 import { GridCache } from './GridCache';
 import { BaseRenderLayer } from './BaseRenderLayer';
-import { CellData, AttributeData, Content } from 'common/buffer/BufferLine';
+import { AttributeData, Content } from 'common/buffer/BufferLine';
 import { NULL_CELL_CODE } from 'common/buffer/Constants';
 import { JoinedCellData } from './CharacterJoinerRegistry';
 import { IColorSet } from 'browser/Types';
+import { CellData } from 'common/buffer/CellData';
 
 /**
  * This CharData looks like a null character, which will forc a clear and render
